deleted gameModel folder and moved contents
